Refactor project structure to follow Python module conventions
