+++ .github/workflows/build_release.yml
@@ -6,26 +6,6 @@ on:
     tags: ['v*']
 
 jobs:
-  build_wheels:
-    name: Build wheels on ${{ matrix.os }}
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        os: [ubuntu-latest, windows-latest, macos-latest]
-
-    steps:
-      - uses: actions/checkout@v4
-
-      - name: Build wheels
-        uses: pypa/cibuildwheel@v2.21.3
-        env:
-          CIBW_BUILD: cp38-* cp39-* cp310-* cp311-*
-          CIBW_SKIP: cp*_i686
-
-      - uses: actions/upload-artifact@v4
-        with:
-          name: smt-${{ matrix.os }}
-          path: ./wheelhouse/*.whl
 
   build_sdist:
     name: Build source distribution
@@ -38,11 +18,11 @@ jobs:
 
       - uses: actions/upload-artifact@v4
         with:
-          name: smt-src
+          name: smt-design-space-ext-src
           path: dist/*.tar.gz
 
   upload_pypi:
-    needs: [build_wheels, build_sdist]
+    needs: [build_sdist]
     runs-on: ubuntu-latest
     environment: pypi
     permissions:
@@ -50,8 +30,6 @@ jobs:
     steps:
       - uses: actions/download-artifact@v4
         with:
-          pattern: smt-*
-          merge-multiple: true
           path: dist
 
       - uses: pypa/gh-action-pypi-publish@release/v1
