+++ requirements.txt
@@ -1,16 +1,5 @@
-Cython
-numpy
-scipy
-scikit-learn
-pyDOE3
-numba           # JIT compiler
-matplotlib      # used in examples and tests
-pytest          # tests runner
-pytest-xdist    # allows running parallel testing with pytest -n <num_workers>
-pytest-cov      # allows to get coverage report
-ruff            # format and lint code
 smt
-jenn >= 1.0.2, <2.0
-egobox ~= 0.20.0
 ConfigSpace==0.6.1 
 adsg-core==1.1.1
+pytest # tests runner
+ruff   # format and lint code
