+++ smt_design_space_ext/tests/test_design_space.py
@@ -7,10 +7,11 @@ import itertools
 import unittest
 
 import numpy as np
-from smt_design_space import design_space as ds
+from smt_design_space_ext import design_space as ds
 
 from smt.sampling_methods import LHS
-from smt_design_space.design_space import (
+
+from smt_design_space_ext.design_space import (
     HAS_CONFIG_SPACE,
     HAS_ADSG,
     DesignSpaceGraph,
@@ -693,7 +694,7 @@ class Test(unittest.TestCase):
     )
     def test_adsg_to_legacy(self):
         from adsg_core import BasicADSG, NamedNode, DesignVariableNode
-        from smt_design_space.design_space import ensure_design_space
+        from smt_design_space_ext.design_space import ensure_design_space
         from adsg_core import GraphProcessor
 
         # Create the ADSG
